fix build errors and rename serviceBusSenderClient
